Switch to int32 type for xsd int fields
The xsd 'int' type is a signed 32-bit integer, whereas 'int' type in Go
is architecture dependent (32 or 64 bits).  Switch to the architecture
independent 'int32' type, which is always 32 bits.

Fixes #344
See also PR #476

+++ govc/dvs/portgroup/info.go
@@ -146,7 +146,7 @@ func (cmd *info) Run(ctx context.Context, f *flag.FlagSet) error {
 		portVlanID := portVlan.VlanId
 
 		// Show port info if: VLAN ID is not defined, or VLAN ID matches requested VLAN
-		if cmd.vlanID == 0 || portVlanID == cmd.vlanID {
+		if cmd.vlanID == 0 || portVlanID == int32(cmd.vlanID) {
 			returnedPorts++
 
 			fmt.Printf("PortgroupKey: %s\n", port.PortgroupKey)
